rename to pr-guide-suggestions
