+++ .github/workflows/pr-guide-suggestions.yaml
@@ -1,7 +1,7 @@
 # Copyright IBM Corp. 2014, 2025
 # SPDX-License-Identifier: MPL-2.0
 ---
-name: Schema Lint
+name: PR Guide Suggestions
 
 permissions:
   contents: read
@@ -11,16 +11,16 @@ on:
   pull_request:
     types: ['opened', 'synchronize']
     paths:
-      - '.github/workflows/schema-lint.yaml'
-      - 'internal/tools/schema-lint/**'
-      - 'scripts/run-schema-lint.sh'
-      - 'scripts/run-schema-lint-diff.sh'
+      - '.github/workflows/pr-guide-suggestions.yaml'
+      - 'internal/tools/pr-guide-suggestions/**'
+      - 'scripts/run-pr-guide-suggestions.sh'
+      - 'scripts/run-pr-guide-suggestions-diff.sh'
       - 'internal/services/**/*_resource.go'
       - 'internal/services/**/*_data_source.go'
       - 'internal/services/**/*_datasource.go'
 
 concurrency:
-  group: 'schema-lint-${{ github.head_ref }}'
+  group: 'pr-guide-suggestions-${{ github.head_ref }}'
   cancel-in-progress: true
 
 jobs:
@@ -36,17 +36,17 @@ jobs:
       - uses: actions/setup-go@4a3601121dd01d1626a1e23e37211e3254c1c06c # v6
         with:
           go-version-file: ./.go-version
-      - name: Run schema lint
+      - name: Run pr-guide-suggestions
         id: lint
         shell: bash
         run: |
           set +e
-          bash ./scripts/run-schema-lint-diff.sh > schema-lint-output.txt 2>&1
+          bash ./scripts/run-pr-guide-suggestions-diff.sh > pr-guide-suggestions-output.txt 2>&1
           code=$?
-          cat schema-lint-output.txt
+          cat pr-guide-suggestions-output.txt
           echo "exit_code=${code}" >> "$GITHUB_OUTPUT"
           exit 0
-      - name: Build schema lint comment
+      - name: Build pr-guide-suggestions comment
         if: ${{ always() && steps.lint.outcome == 'success' }}
         shell: bash
         env:
@@ -55,7 +55,7 @@ jobs:
           BASE_REF: ${{ github.base_ref }}
           RUN_URL: ${{ github.server_url }}/${{ github.repository }}/actions/runs/${{ github.run_id }}
         run: |
-          output=$(sed -E 's/\x1b\[[0-9;]*m//g' schema-lint-output.txt \
+          output=$(sed -E 's/\x1b\[[0-9;]*m//g' pr-guide-suggestions-output.txt \
             | sed -e 's/&/\&amp;/g' -e 's/</\&lt;/g' -e 's/>/\&gt;/g')
           if [ -z "$(printf '%s' "$output" | tr -d '[:space:]')" ]; then
             output="No newly added schema properties required linting."
@@ -65,10 +65,10 @@ jobs:
             output="${output:0:$max}"$'\n''... (truncated, see the workflow run for the full output)'
           fi
           if [ "$EXIT_CODE" = "0" ]; then
-            heading="### 🧹 Schema Lint — ✅ Passed"
+            heading="### 🧹 PR Guide Suggestions — ✅ Passed"
             details_open=""
           else
-            heading="### 🧹 Schema Lint — ❌ Failed"
+            heading="### 🧹 PR Guide Suggestions — ❌ Failed"
             details_open=" open"
           fi
           {
@@ -81,27 +81,27 @@ jobs:
             echo "</details>"
             if [ "$EXIT_CODE" != "0" ]; then
               echo ""
-              echo "> One or more newly added schema properties violate a schema-lint rule. Correct the property (auto-fixable naming findings can be applied with the linter's \`-fix\` flag); if the change is intentional, add a \`//nolint\` (or \`//nolint:<RULE>\`) comment to the property, or a maintainer can apply the \`schema-accepted\` label to bypass this check."
+              echo "> One or more newly added schema properties violate a pr-guide-suggestions rule. Correct the property (auto-fixable naming findings can be applied with the linter's \`-fix\` flag); if the change is intentional, add a \`//nolint\` (or \`//nolint:<RULE>\`) comment to the property, or a maintainer can apply the \`schema-accepted\` label to bypass this check."
               echo ">"
-              echo "> Reproduce locally: \`GITHUB_BASE_REF=${BASE_REF} bash ./scripts/run-schema-lint-diff.sh\`"
+              echo "> Reproduce locally: \`GITHUB_BASE_REF=${BASE_REF} bash ./scripts/run-pr-guide-suggestions-diff.sh\`"
             fi
-          } > schema-lint-comment.md
-      - name: Comment schema lint results
+          } > pr-guide-suggestions-comment.md
+      - name: Comment pr-guide-suggestions results
         if: ${{ always() && steps.lint.outcome == 'success' }}
         continue-on-error: true
         uses: marocchino/sticky-pull-request-comment@5770ad5eb8f42dd2c4f34da00c94c5381e49af88 # v3.0.5
         with:
-          header: schema-lint
-          path: schema-lint-comment.md
-      - name: Fail if schema lint reported findings
+          header: pr-guide-suggestions
+          path: pr-guide-suggestions-comment.md
+      - name: Fail if pr-guide-suggestions reported findings
         if: ${{ steps.lint.outputs.exit_code != '0' }}
         run: |
-          echo "::error::Schema lint failed for a newly added schema property."
+          echo "::error::PR guide suggestions failed for a newly added schema property."
           echo ""
-          echo "This PR adds one or more schema properties that violate a schema-lint"
-          echo "rule. See the schema-lint comment on the pull request for details, or"
+          echo "This PR adds one or more schema properties that violate a pr-guide-suggestions"
+          echo "rule. See the pr-guide-suggestions comment on the pull request for details, or"
           echo "reproduce locally against the branch you are merging into:"
-          echo "  GITHUB_BASE_REF=${GITHUB_BASE_REF:-main} bash ./scripts/run-schema-lint-diff.sh"
+          echo "  GITHUB_BASE_REF=${GITHUB_BASE_REF:-main} bash ./scripts/run-pr-guide-suggestions-diff.sh"
           echo ""
-          echo "See internal/tools/schema-lint/README.md for the full rule list."
+          echo "See internal/tools/pr-guide-suggestions/README.md for the full rule list."
           exit 1
