+++ internal/tools/pr-guide-suggestions/lint/lint_test.go
@@ -8,7 +8,7 @@ import (
 	"path/filepath"
 	"testing"
 
-	"github.com/hashicorp/terraform-provider-azurerm/internal/tools/schema-lint/rules"
+	"github.com/hashicorp/terraform-provider-azurerm/internal/tools/pr-guide-suggestions/rules"
 )
 
 // lintSrc writes src to a temporary *_resource.go file and lints it. Because the
