+++ internal/tools/pr-guide-suggestions/rules/resolution_test.go
