+++ internal/tools/pr-guide-suggestions/rules/sl006_test.go
